fix: remove nav item constatn
